+++ functions.php
@@ -21,7 +21,7 @@ require_once 'inc/_conf/setup-theme.php';
 require_once 'inc/_conf/setup-editor.php';
 require_once 'inc/_conf/register-assets.php';
 // require_once 'inc/_conf/register-menus.php'; // Uncomment his if you wan't to use Classic menus. You need also a custom block or Avidly Classic Menu block to display classic menus from backend.
-require_once 'inc/_conf/register-block-patterns.php';
+require_once 'inc/_conf/register-pattern-categories.php';
 require_once 'inc/_conf/register-block-templates.php';
 require_once 'inc/_conf/register-block-styles.php';
 
@@ -32,6 +32,7 @@ require_once 'inc/helpers/theme-hooks.php';
 require_once 'inc/helpers/override-core-styles.php';
 require_once 'inc/helpers/featured-images.php';
 require_once 'inc/helpers/helper-functions.php';
+require_once 'inc/helpers/helper-patterns.php';
 
 /**
  * Block queries