+++ patterns/_template-01-page.php
@@ -1,0 +1,18 @@
+<?php
+/**
+ * Title: Sisältösivu: oletus
+ * Description: Page template for content page, default.
+ * Slug: avidly-theme/template-01-page
+ * viewportWidth: 1440
+ * Block Types: core/post-content
+ * Post Types: page, wp_template
+ * Categories: template
+ *
+ * @package Avidly_Theme
+ */
+
+?>
+
+<!-- wp:pattern {"slug":"avidly-theme/hidden-hero-01"} /-->
+
+<!-- wp:pattern {"slug":"avidly-theme/content-01-columns-bg"} /-->