+++ patterns/_template-02-page.php
@@ -1,0 +1,32 @@
+<?php
+/**
+ * Title: Sisältösivu: yksinkertainen
+ * Description: Page template for content page, simple.
+ * Slug: avidly-theme/template-02-page
+ * viewportWidth: 1440
+ * Block Types: core/post-content
+ * Post Types: page, wp_template
+ * Categories: template
+ *
+ * @package Avidly_Theme
+ */
+
+?>
+
+<!-- wp:pattern {"slug":"avidly-theme/hidden-hero-02"} /-->
+
+<!-- wp:spacer {"height":"5rem"} -->
+<div style="height:5rem" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->
+
+<!-- wp:paragraph {"placeholder":"Lisää ingressi tähän.","fontSize":"medium"} -->
+<p class="has-medium-font-size"></p>
+<!-- /wp:paragraph -->
+
+<!-- wp:paragraph -->
+<p></p>
+<!-- /wp:paragraph -->
+
+<!-- wp:spacer {"height":"5rem"} -->
+<div style="height:5rem" aria-hidden="true" class="wp-block-spacer"></div>
+<!-- /wp:spacer -->