+++ patterns/content-01-columns-bg.php
@@ -1,24 +1,23 @@
 <?php
 /**
- * Title: Three columns with background color
- * Description: Add block description.
- * Slug: avidly-theme/post-hero
+ * Title: Osio taustavärillä, kolme saraketta.
+ * Description: Editing the content of the Columns block is restricted ("templateLock":"contentOnly").
+ * Slug: avidly-theme/content-01-columns-bg
  * viewportWidth: 1440
  * Block Types: core/post-content
- * Post Types: page, post
  * Categories: content
  *
  * @package Avidly_Theme
  */
 
 ?>
-<!-- wp:group {"align":"full","backgroundColor":"lightgray","layout":{"type":"constrained"}} -->
-<div class="wp-block-group alignfull has-lightgray-background-color has-background">
+<!-- wp:group {"align":"full","backgroundColor":"lightgray","layout":{"type":"constrained"},"className":"pattern-content-01-columns-bg"} -->
+<div class="wp-block-group alignfull has-lightgray-background-color has-background pattern-content-01-columns-bg">
 	<!-- wp:heading {"textAlign":"center"} -->
 	<h2 class="wp-block-heading has-text-align-center">Otsikko h2</h2>
 	<!-- /wp:heading -->
 
-	<!-- wp:columns {"align":"wide"} -->
+	<!-- wp:columns {"templateLock":"contentOnly", "align":"wide"} -->
 	<div class="wp-block-columns alignwide"><!-- wp:column -->
 		<div class="wp-block-column">
 		<?php