+++ patterns/hidden-hero-02.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * Title: Hero-yliö, yksinkertainen (piilotettu)
+ * Description: Pattern is added within another block pattern, therefore it is hidden from the 'Patterns' menu (Inserter: no). Editing the content of the Group block is restricted ("templateLock":"contentOnly").
+ * Slug: avidly-theme/hidden-hero-02
+ * viewportWidth: 1440
+ * Block Types: core/post-content
+ * Inserter: no
+ * Categories: hero
+ *
+ * @package Avidly_Theme
+ */
+
+?>
+<!-- wp:group {"templateLock":"contentOnly","align":"wide","style":{"spacing":{"margin":{"top":"var:preset|spacing|xl","bottom":"var:preset|spacing|xl"}}},"layout":{"type":"default"},"className":"pattern-hidden-hero-02"} -->
+<div class="wp-block-group alignwide pattern-hidden-hero-02" style="margin-top:var(--wp--preset--spacing--xl);margin-bottom:var(--wp--preset--spacing--xl)">
+	<!-- wp:post-title {"textAlign":"center","level":1} /-->
+</div>
+<!-- /wp:group -->